api/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.22
google-api-python-client>=2.129.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-generativeai>=0.8.0
pypdf>=4.2.0
